Add HTTPWithResult task
This lets us use the celery worker to submit jobs to the shell-command
service and wait for it to complete to give us the job_id.  We were just
doing it in the webserver, which didn't let us share the code we
already wrote to retry connection errors and such.