Set order of top-level links list
This enables testing by returning the list of workflow links in a
predictable order.  I'm concerned that adding this sorting could be a
potential liability in the future while processing requests involving
workflows with many links.

One alternative to returning lists of links in a canonical order to
support testing would be to use a library for comparing data that
supports comparing data as "bag", like in Test::Deep from CPAN.