Add Executions both before and after the split
This is for parallel-by Tasks.  By creating Executions after the split
(as well as update actions for those Executions), we'll better match the
Executions on the Methods under these Tasks.

Previous to this, only the parent-color'd Execution was created.  For
instance:

Before:
    If ||-by (a, b, c) we'd just get one Execution with color=0

Now:
    we get one Execution with color=0 and another 3 with color=1,2,3
    that each have parent-color=0.

The Methods under a Task that is Parallel by will have 3 Executions with
color=1,2,3 that each have parent-color=0. Importantly, they will NOT have the
Execution with color=0.