flatten query-string args
request.args is of class ImmutableMultiDict, and by default, when it is
cast into a python dict is to make every value a list, even if they are
only one element long.  Presumably this is to support query string
arguments that are repeated and intended to become lists.  We are not
using such a feature, so flat=True ensures that the values of are
scalars.