Changed JSON schema for links
The new representation is less repetitive in many cases.  It also is
more readable since the link section isn't a flat structure of similar
items anymore.  The real reason to change it though is that we plan to
normalize the link database representation in a similar way while
implementing support for control-flow only links.  So while the
'dataFlow' section is required now, it will become optional.