Allow specifying webhooks for job services
PREVIOUSLY: even though you would specify the webhooks in the parameters
section of the method, they would not be sent to the job service.
Instead, only the webhooks needed by the workflow service would be sent
to the job service.  The webhooks specified in the parameters section
would be sent from the workflow service instead.

NOW: You may specify webhooks either on the method or in the parameters
section.  If specified on the method, they will be sent from the
workflow service, otherwise they will be sent from the job service.