Enable ON DELETE CASCADEs
These are postgres server side directives on the foreignkey constraint.
Sqlalchemy has to be told to not try to set the foreignkey to null, to
do this we use the passive_deletes='all' option on the defined
relationships.

To migrate the database, we have to delete the existing foreignkey
constraints and add them back with the ON DELETE CASCADE directive.