Add 'limited-workflow-executions' report
This report supports the following query string args:

workflow_id
since => returns only executions with timestamps at or after this
limit => returns at most this many results

The response includes:

executions => a list of executions
updateUrl => a url to fetch more executions (with 'since' query string
argument already specified appropriately*)
numRemaining => number of executions (currently) beyond the limit

* The way 'since' is calculated for 'updateUrl' is to use the timestamp
of the first execution outside of the limit.  That is, if we're
restricted to returning 10 executions, we use the timestamp of the
11th execution for the 'since' query string argument.  This has the
nice effect of reducing (often to zero) the number of executions that
are sent in multiple responses.  If this timestamp is coincidentally the
same as the timestamp of the 10th execution (for example) then no
reduction in duplication occurs.

If a large batch of executions have exactly the same timestamp, the same
set of executions could be returned over and over, I see this as
unlikely and something that could be easily detected and resolved (by
increasing the 'limit' value).