Break chain
The Chain:

When the petri service tells the workflow service to execute a job
method, the workflow service doesn't respond until it has recieved a
response from the job service.

How its broken:

The workflow service now responds just after calling .delay() on a
celery task that will submit the job to the job service and ensure
'errored' status is set if it cannot.